[REFACTOR]: SearchBar 폴더구조 수정
