+++ components/SearchBar/ContainerStyle.ts
@@ -1,12 +1,12 @@
 import styled from '@emotion/styled';
 import { VerticalAlign } from '@styles/common';
 
-export const SearchBarContainer = styled.div`
+export const Container = styled.div`
   position: relative;
   z-index: 1;
 `;
 
-export const SearchBarContentContainer = styled.div`
+export const ContentContainer = styled.div`
   width: 100%;
   height: 50%;
   ${VerticalAlign};
